requirements.txt
pyyaml
click
chalk